+++ src/grid/grid-filtering.component.ts
@@ -106,6 +106,7 @@ export class IgxGridFilterComponent implements IGridBus, OnInit, OnDestroy, DoCh
         return `igx-filtering`;
     }
 
+    public booleanFilterAll = "All";
     public dialogShowing = false;
     public dialogPosition = "igx-filtering__options--to-right";
 
@@ -133,6 +134,9 @@ export class IgxGridFilterComponent implements IGridBus, OnInit, OnDestroy, DoCh
 
     @ViewChild("select", { read: ElementRef})
     protected select: ElementRef;
+
+    @ViewChild("input", { read: ElementRef})
+    protected input: ElementRef;
 
     constructor(private zone: NgZone, public gridAPI: IgxGridAPIService, public cdr: ChangeDetectorRef, private elementRef: ElementRef) {
         this.filterChanged.pipe(
@@ -167,6 +171,7 @@ export class IgxGridFilterComponent implements IGridBus, OnInit, OnDestroy, DoCh
     public refresh() {
         this.dialogShowing = !this.dialogShowing;
         if (this.dialogShowing) {
+            this.focusInput();
             this.column.filteringCondition = this.getCondition(this.select.nativeElement.value);
         }
         this.cdr.detectChanges();
@@ -210,9 +215,15 @@ export class IgxGridFilterComponent implements IGridBus, OnInit, OnDestroy, DoCh
         if (this.dataType === DataType.Date) {
             this.cdr.detectChanges();
         }
+        this.focusInput();
     }
 
     public selectionChanged(value): void {
+        if (value === this.booleanFilterAll) {
+            this.clearFiltering(true);
+            return;
+        }
+        this.focusInput();
         this._filterCondition = value;
         this.column.filteringCondition = this.getCondition(value);
         if (this.unaryCondition) {
@@ -230,10 +241,15 @@ export class IgxGridFilterComponent implements IGridBus, OnInit, OnDestroy, DoCh
         this.clearFiltering(false);
     }
 
+    public focusInput(): void {
+        if (this.input) {
+            this.input.nativeElement.focus();
+        }
+    }
+
     public get disabled() {
-        if ((!!this.value || this.value === 0) && !this.unaryCondition) {
-            return false;
-        } else if (this.unaryCondition) {
+        // if filtering is applied, reset button should be active
+        if (this.gridAPI.get(this.gridID).filteringExpressions.length > 0) {
             return false;
         }
         return true;