+++ src/grid/index.ts
@@ -8,6 +8,7 @@ import { IgxButtonModule } from "../directives/button/button.directive";
 import { IgxFocusModule } from "../directives/focus/focus.directive";
 import { IgxForOfModule } from "../directives/for-of/for_of.directive";
 import { IgxRippleModule } from "../directives/ripple/ripple.directive";
+import { IgxTextHighlightModule } from "../directives/text-highlight/text-highlight.directive";
 import { IgxTextSelectionModule } from "../directives/text-selection/text-selection.directive";
 import { IgxToggleModule } from "../directives/toggle/toggle.directive";
 import { IgxIconModule } from "../icon";
@@ -76,6 +77,7 @@ import { IgxGridRowComponent } from "./row.component";
     IgxToggleModule,
     IgxForOfModule,
     IgxFocusModule,
+    IgxTextHighlightModule,
     IgxTextSelectionModule,
     IgxCheckboxModule
   ],